refactor: separate Hyperjump and inferenceai code and layout
