+++ app/[lang]/(hyperjump)/case-studies/[slug]/components/content.tsx
