+++ app/[lang]/(hyperjump)/case-studies/data.ts
