+++ app/[lang]/(hyperjump)/jobs/[id]/page.tsx
