+++ app/[lang]/(hyperjump)/services/page.tsx
@@ -1,8 +1,11 @@
 import Link from "next/link";
 import Image from "next/image";
 
-import { Clients } from "@/app/components/clients";
-import { serviceBySlug, services, ServiceSlug } from "@/app/data";
+import {
+  serviceBySlug,
+  services,
+  ServiceSlug
+} from "@/app/[lang]/(hyperjump)/data";
 import { Button } from "@/components/ui/button";
 import data from "@/data.json";
 import { cn } from "@/lib/utils";
@@ -19,6 +22,8 @@ import {
   servicesHeroHeading,
   servicesHeroDesc
 } from "@/locales/.generated/server";
+
+import { Clients } from "../components/clients";
 
 export const generateStaticParams = async () => {
   return supportedLanguages.map((lang) => ({ lang }));
