+++ app/[lang]/(hyperjump)/smdd2024/components/case-studies.tsx
