+++ app/[lang]/(hyperjump)/smdd2024/components/hero.tsx
