+++ app/[lang]/(hyperjump)/smdd2024/components/projects.tsx
