+++ app/[lang]/(hyperjump)/smdd2024/components/trusted.tsx
