+++ app/[lang]/(inferenceai)/inferenceai/[slug]/data.tsx
