+++ app/[lang]/(inferenceai)/inferenceai/[slug]/home.tsx
