+++ app/[lang]/(inferenceai)/inferenceai/[slug]/page.tsx
