+++ app/[lang]/(inferenceai)/inferenceai/components/client-wrapper.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import { usePathname } from "next/navigation";
+import type { SupportedLanguage } from "@/locales/.generated/types";
+import Nav from "./nav";
+import { navRagChatbot } from "../[slug]/data";
+import { navInferenceai } from "../data";
+
+type ClientWrapperProps = {
+  lang: SupportedLanguage;
+};
+
+export default function ClientWrapper({ lang }: ClientWrapperProps) {
+  const pathname = usePathname();
+  const isInferenceAiCaseStudy = pathname.includes(`/${lang}/inferenceai/`);
+
+  return (
+    <Nav
+      menus={
+        isInferenceAiCaseStudy ? navRagChatbot(lang) : navInferenceai(lang)
+      }
+      lang={lang}
+    />
+  );
+}
