+++ app/[lang]/(inferenceai)/inferenceai/home.tsx
