+++ app/[lang]/(inferenceai)/inferenceai/page.tsx
