--- app/styles/ai-agent.css
@@ -1,81 +1,0 @@
-/** Chat UI **/
-@import "tailwindcss";
-
-/** Header **/
-#n8n-chat .chat-header {
-  @apply gap-2 p-6;
-}
-
-#n8n-chat .chat-heading {
-  @apply leading-none font-bold;
-}
-
-/** Chat **/
-#n8n-chat .chat-window {
-  @apply rounded-xl border-none shadow-lg transition-all duration-300;
-}
-
-#n8n-chat .chat-message-markdown {
-  @apply text-base;
-}
-
-#n8n-chat .chat-message {
-  @apply rounded-md;
-}
-
-/** Footer **/
-#n8n-chat .chat-footer {
-  @apply bg-white p-2;
-}
-
-#n8n-chat .chat-powered-by {
-  @apply text-[0px];
-}
-
-#n8n-chat .chat-powered-by {
-  --custom-content: "HyperBot made with \2665  by Hyperjump";
-  @apply relative text-transparent before:inline-block before:text-sm before:text-black before:content-(--custom-content);
-}
-
-#n8n-chat .chat-powered-by a {
-  @apply hidden;
-}
-
-/** Floating action button **/
-#n8n-chat .chat-window-toggle {
-  @apply bg-[#3276F5];
-}
-
-@media screen and (min-width: 992px) {
-  #n8n-chat .chat-window-toggle {
-    @apply mx-auto !hidden;
-  }
-
-  #n8n-chat .chat-header {
-    @apply relative;
-  }
-
-  #n8n-chat .chat-window-wrapper.chat-window-minimized {
-    @apply -z-10 mr-8 opacity-0;
-  }
-
-  #n8n-chat .chat-window-wrapper {
-    @apply bottom-0 mx-auto mt-8 h-[600px] w-full transition-all duration-300;
-  }
-
-  #n8n-chat .chat-window-wrapper.chat-window-centered {
-    @apply left-0 mx-auto max-w-4xl;
-  }
-
-  #n8n-chat .chat-window-wrapper.chat-window-right {
-    @apply right-0 mx-auto mr-4 max-w-lg;
-  }
-
-  #n8n-chat .chat-window {
-    @apply w-full;
-  }
-}
-
-#n8n-chat .chat-window-toggle {
-  @apply bg-[#3276F5];
-}
